requirements.txt
streamlit>=1.32.0
pytesseract>=0.3.10
Pillow>=10.0.0
opencv-python-headless>=4.9.0
requests>=2.31.0
python-dotenv>=1.0.0
google-generativeai>=0.5.0
groq>=0.9.0
ollama>=0.2.0
numpy>=1.24.0
reportlab>=4.0.0
pdf2image>=1.16.0
pandas>=1.4.0
openpyxl>=3.1.0
